requirements.txt
ffmpeg-python
openai-whisper
nltk 